+++ CHANGES.md
@@ -1,3 +1,9 @@
+### Version 0.0.3
+- [FEATURE] Create Frontend Plugin and add it to PAGE, when popup is enabled
+- [BUGFIX] Renamed label in backend module for hidden field
+- [WIP] Base Appearance for Popup
+- [WIP] Set page id with Popup content
+
 ### Version 0.0.2
 - [IMPORTANT] Add Backend Module to create and remove configuration
 - [FEATURE] Configuration model to enable popup for pages